+++ Assets/Features/Global/Services/Network/Objects/Factories/Abstract/IEntityFactory.cs
@@ -5,9 +5,9 @@ namespace Global.Network.Objects.Factories.Abstract
 {
     public interface IEntityFactory
     {
-        int Id { get; }
+        ushort Id { get; }
 
         UniTaskVoid OnRemoteCreated(int objectId, RagonEntity entity);
-        void AssignId(int id);
+        void AssignId(ushort id);
     }
 }
